NodeJS e2e fixes
