+++ src/execution_service/errors/InvalidEndpoint.ts
@@ -1,0 +1,8 @@
+import { CustomError } from 'ts-custom-error'
+
+export class InvalidEndpoint extends CustomError {
+  public constructor (validEndpoints: string[]) {
+    super()
+    this.message = `Endpoint does not exist. Options are: ${validEndpoints}`
+  }
+}
