+++ src/execution_service/errors/index.ts
@@ -3,3 +3,4 @@ export { BadArgument } from './BadArgument'
 export { ContractNotLoaded } from './ContractNotLoaded'
 export { MissingConfig } from './MissingConfig'
 export { ContainerFailedToStart } from './ContainerFailedToStart'
+export { InvalidEndpoint } from './InvalidEndpoint'
