+++ src/server/infrastructure/engine_clients/StubEngineClient.ts
@@ -12,14 +12,16 @@ export function StubEngineClient (): EngineClient {
     call ({ contractAddress, method, methodArguments }: ContractCallInstruction) {
       if (method === 'schema') {
         return {
-          schema: `
-            const addArgs = t.type({
-              number1: t.number,
-              number2: t.number,
-            })
+          data: {
+            data: `
+              const addArgs = t.type({
+                number1: t.number,
+                number2: t.number,
+              })
 
-            export const Add = createEndpoint<typeof addArgs, typeof t.number, t.NullC>('Add', addArgs, t.number)
-          `
+              export const Add = createEndpoint<typeof addArgs, typeof t.number, t.NullC>('Add', addArgs, t.number)
+            `
+          }
         }
       }
