+++ test/bundles/nodejs/abcd
@@ -7,7 +7,7 @@
         number2: t.number,
       })
 
-      export const Add = createEndpoint<typeof addArgs, typeof t.number, t.NullC>('Add', addArgs, t.number)
+      export const add = createEndpoint<typeof addArgs, typeof t.number, t.NullC>('add', addArgs, t.number)
     `
   }
 }
