--- test/e2e/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2015",
-    "noEmitOnError": true,
-    "noImplicitAny": true,
-    "noUnusedParameters": true,
-    "noUnusedLocals": true,
-    "removeComments": true,
-    "preserveConstEnums": true,
-    "outDir": "dist/",
-    "sourceMap": true,
-    "composite": true,
-    "rootDir": "../..",
-    "allowSyntheticDefaultImports": true
-  },
-  "include": [
-    "src/**/*",
-    "../../src/client/**/*"
-  ]
-}
