--- test/smart_contract_server_mock/.dockerignore
@@ -1,1 +1,0 @@
-target
