--- test/smart_contract_server_mock/.gitignore
@@ -1,1 +1,0 @@
-target
