--- test/smart_contract_server_mock/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "smart_contract_server"
-version = "0.1.0"
-authors = ["sam <sam.jeston@iohk.io>"]
-edition = "2018"
-
-[dependencies]
-rocket = "0.4.0"
-serde = "1.0"
-serde_json = "1.0"
-serde_derive = "1.0"
-
-[dependencies.rocket_contrib]
-version = "*"
-default-features = false
-features = ["json"]
