--- test/smart_contract_server_mock/Rocket.toml
@@ -1,7 +1,0 @@
-[development]
-address = "localhost"
-port = 8080
-
-[production]
-address = "0.0.0.0"
-port = 8080
